refactor: Declare employees as enum
