+++ src/main/java/domain/Employee.java
@@ -1,12 +1,10 @@
 package domain;
 
-import java.util.Date;
+import enums.EmployeeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,17 +17,16 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AcceptancePolicy {
+public class Employee {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private int id;
 
-  private Date createdDate;
-  private String description;
+  private EmployeeType employeeType;
+  private int age;
+  private String email;
   private String name;
-
-  @OneToOne
-  @JoinColumn(name = "employee_id")
-  private Employee writer;
+  private String phoneNumber;
+  private boolean sex;
 }
