--- src/main/java/domain/employee/Actuary.java
@@ -1,24 +1,0 @@
-package domain.employee;
-
-public class Actuary extends Employee {
-
-  public Actuary() {
-
-  }
-
-  public void designProduct() {
-
-  }
-
-  public void findProduct(String productCode) {
-
-  }
-
-  public void manageProduct(String productCode) {
-
-  }
-
-  public void requestAuthorization(String productCode) {
-
-  }
-}
