--- src/main/java/domain/employee/Sales.java
@@ -1,34 +1,0 @@
-package domain.employee;
-
-public class Sales extends Employee {
-
-  private boolean admin;
-
-  public Sales() {
-
-  }
-
-  public boolean isAdmin() {
-    return admin;
-  }
-
-  public void setAdmin(boolean admin) {
-    this.admin = admin;
-  }
-
-  public void createContract() {
-
-  }
-
-  public void manageCustomer() {
-
-  }
-
-  public void manageOrganization() {
-
-  }
-
-  public void manageProduct() {
-
-  }
-}
