--- src/main/java/domain/employee/Underwriter.java
@@ -1,28 +1,0 @@
-package domain.employee;
-
-public class Underwriter extends Employee {
-
-  public Underwriter() {
-
-  }
-
-  public void createAcceptancePolicy() {
-
-  }
-
-  public void manageCollaboration() {
-
-  }
-
-  public void manageReinsurance() {
-
-  }
-
-  public void mangeLossRate() {
-
-  }
-
-  public void underwrite() {
-
-  }
-}
