--- src/main/java/domain/employee/compensation/AccidentInvestigator.java
@@ -1,16 +1,0 @@
-package domain.employee.compensation;
-
-public class AccidentInvestigator extends Compensation {
-
-  public AccidentInvestigator() {
-
-  }
-
-  public void dispatch() {
-
-  }
-
-  public void receiveDispatch() {
-
-  }
-}
