--- src/main/java/domain/employee/compensation/Compensation.java
@@ -1,10 +1,0 @@
-package domain.employee.compensation;
-
-import domain.employee.Employee;
-
-public class Compensation extends Employee {
-
-  public Compensation() {
-
-  }
-}
