--- src/main/java/domain/employee/compensation/CustomerServiceAgent.java
@@ -1,16 +1,0 @@
-package domain.employee.compensation;
-
-public class CustomerServiceAgent extends Compensation {
-
-  public CustomerServiceAgent() {
-
-  }
-
-  public void counsel() {
-
-  }
-
-  public void receiveAccident() {
-
-  }
-}
