+++ src/main/java/enums/EmployeeType.java
@@ -1,0 +1,10 @@
+package enums;
+
+public enum EmployeeType {
+  ACCIDENT_INVESTIGATOR,
+  ADJUSTER,
+  CUSTOMER_SERVICE_AGENT,
+  ACTUARY,
+  SALES,
+  UNDERWRITER
+}
